auto-leasing-doc-issuance-1.0-main/auto-leasing-doc-issuance-1.0-main/requirements.txt
streamlit
python-docx
babel
num2words
Pillow
